Decouple channel API from exchange manager using delegate (#6237)
* Decouple channel API from exchange manager using delegate

* address review comments
