+++ src/channel/tests/BUILD.gn
@@ -1,4 +1,4 @@
-# Copyright (c) 2020 Project CHIP Authors
+# Copyright (c) 2021 Project CHIP Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,15 +20,9 @@ import("//build_overrides/nlunit_test.gni")
 import("${chip_root}/build/chip/chip_test_suite.gni")
 
 chip_test_suite("tests") {
-  output_name = "libMessagingLayerTests"
+  output_name = "libChannelAPITests"
 
-  sources = [
-    "MessagingContext.cpp",
-    "MessagingContext.h",
-    "TestExchangeMgr.cpp",
-    "TestMessagingLayer.h",
-    "TestReliableMessageProtocol.cpp",
-  ]
+  sources = [ "TestChannel.cpp" ]
 
   cflags = [ "-Wconversion" ]
 
@@ -44,8 +38,5 @@ chip_test_suite("tests") {
     "${nlunit_test_root}:nlunit-test",
   ]
 
-  tests = [
-    "TestExchangeMgr",
-    "TestReliableMessageProtocol",
-  ]
+  tests = [ "TestChannel" ]
 }
